feat: hook to check auth
